# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Choregos API
  version: 0.17.11
  description: |
    The Choregos platform API. The console and the CLI use this API and nothing else.
    Auth: an OIDC session (cookie) or `Authorization: Bearer <api_token>`.
    The `/internal/*` routes are reserved for runners and authenticated by a run JWT (ES256).
    Errors: RFC 9457 (`application/problem+json`). Cursor pagination (`cursor`, `limit`).
  license: { name: Apache-2.0, url: https://www.apache.org/licenses/LICENSE-2.0 }
servers:
  - url: /api/v1
    description: API base
security:
  - session: []
  - bearer: []
...
paths:
...
  /projects/{id}/work-items:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [work-items]
      operationId: listWorkItems
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
        - { name: state, in: query, schema: { type: string } }
        - { name: size, in: query, schema: { type: string, enum: [S, M, L, XL] } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemPage" } } } }
    post:
      tags: [work-items]
      operationId: createWorkItem
      summary: File a request in Choregos (internal tracker only)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemCreate" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /work-items/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [work-items]
      operationId: getWorkItem
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [work-items]
      operationId: updateWorkItem
      summary: Change a ticket's fields
      description: >-
        The given fields replace the ticket's own, `null` removes one; the whole is validated by the
        pinned workflow's `metadata.inputs`. The interpreter is told (`fields_changed`): a moved
        start date re-arms the action waiting for it (S20-05).
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemUpdate" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/WorkItem" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /work-items/{id}/timeline:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [work-items]
      operationId: getWorkItemTimeline
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/TimelineEntry" } } } } }
  /work-items/{id}/hand-offs:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [work-items]
      operationId: getWorkItemHandOffs
      description: >-
        The hand-off receipts of a work item (S25-04), output by output: every run that produced it
        and every run that read it, under the SHA-256 digest of the revision it produced or read, and
        the digest of what the output is worth today — an output modified after it was produced no
        longer has the digest of its last receipt.
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/HandOff" } } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /work-items/{id}/journey:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [work-items]
      operationId: getWorkItemJourney
      description: >-
        The work item's path through its workflow, in one read: the map of the version it is pinned
        to, every move from state to state with the transition that carried it, and every step an
        actor took — agent runs, human requests, governed actions. The console animates it.
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemJourney" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /work-items/{id}/decisions:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [work-items]
      operationId: postDecision
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DecisionRequest" } } }
      responses:
        "202": { description: Decision passed on to the workflow, content: { application/json: { schema: { $ref: "#/components/schemas/HumanRequest" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /work-items/{id}/actions:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [work-items]
      operationId: postWorkItemAction
      description: >-
        `migrate` requires `workflow_def_id`, a definition of the project; the ticket's current
        state must exist in it, or be mapped by `state_mapping` — otherwise 422, and nothing is
        passed on.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/WorkItemAction" } } }
      responses:
        "202": { description: Action passed on }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
...
components:
  securitySchemes:
    session: { type: apiKey, in: cookie, name: choregos_session }
    bearer: { type: http, scheme: bearer, bearerFormat: choregos-api-token }
    runToken: { type: http, scheme: bearer, bearerFormat: JWT, description: "Run JWT signed with ES256, aud=internal" }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
...
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
  responses:
...
    Forbidden:
      description: Forbidden (RBAC)
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    NotFound:
      description: Not found
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Conflict:
      description: State conflict
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    ValidationError:
      description: Unprocessable entity
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
  schemas:
    Problem:
      type: object
      description: RFC 9457
      required: [type, title, status]
      properties:
        type: { type: string, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              loc: { type: array, items: { type: string } }
              msg: { type: string }
              line: { type: integer }
              column: { type: integer }
        retry_after: { type: integer }
    PageMeta:
      type: object
      required: [has_more]
      properties:
        next_cursor: { type: [string, "null"] }
        has_more: { type: boolean }
...
    Size: { type: string, enum: [S, M, L, XL] }
    Risk: { type: string, enum: [low, medium, high] }
...
    ProcessStep:
      type: object
      description: A transition in plain words — the process view (ADR 0031).
      required: [id, from, to, actor, actor_type, who, sentence]
      properties:
        id: { type: string }
        from: { type: string }
        from_display: { type: string }
        to: { type: string }
        to_display: { type: string }
        actor: { type: string }
        actor_type: { type: string, enum: [agent, human, system, release_train] }
        who: { type: string }
        outputs: { type: array, items: { type: string } }
        gates: { type: array, items: { type: object, properties: { name: { type: string }, summary: { type: string } } } }
        on_fail: { type: [string, "null"] }
        on_reject: { type: [string, "null"] }
        timeout_hours: { type: [integer, "null"] }
        sentence: { type: string }
...
    WorkflowGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            type: object
            required: [id, display, kind]
            properties:
              id: { type: string }
              display: { type: string }
              kind: { type: string }
              terminal: { type: boolean }
              lane: { type: string }
              tracker_status: { type: [string, "null"], description: "The tracker column the state maps to; the console groups the path into stages by it." }
        edges:
          type: array
          items:
            type: object
            required: [from, to]
            properties:
              id: { type: [string, "null"] }
              from: { type: string }
              to: { type: string }
              kind:
                type: string
                description: "`nominal` for a transition of the YAML; `reject`, `resume`, `escalate`, `default`… for the secondary edges (drawn dashed)."
              label: { type: string }
              wildcard: { type: boolean }
              actor: { type: [string, "null"] }
              actor_type: { type: [string, "null"] }
              role: { type: [string, "null"] }
              gates: { type: array, items: { type: string } }
              via: { type: [string, "null"] }
              timeout_hours: { type: [integer, "null"] }
...
    WorkItemCreate:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 500 }
        body: { type: string, default: "" }
        size: { type: [string, "null"], enum: [S, M, L, XL, null] }
        risk: { type: [string, "null"], enum: [low, medium, high, null] }
        start: { type: boolean, default: true }
        workflow: { type: [string, "null"], description: "The workflow the ticket starts in; otherwise the routing, otherwise the default one (ADR 0031)." }
        labels: { type: array, items: { type: string }, description: "Read by the routing rules." }
        fields: { type: object, additionalProperties: true, description: "The ticket's fields, validated by its workflow's `metadata.inputs`." }
    WorkItemUpdate:
      type: object
      required: [fields]
      additionalProperties: false
      properties:
        fields: { type: object, additionalProperties: true, description: "The fields that change; `null` removes one." }
...
    WorkflowFailure:
      type: object
      description: Why a ticket's interpreter died. The ticket will not move again without a restart.
      required: [message]
      properties:
        message: { type: string }
        activity: { type: [string, "null"] }
        state: { type: [string, "null"] }
        at: { type: [string, "null"], format: date-time }
    WorkItem:
      type: object
      required: [id, project_slug, tracker_key, title, state, totals]
      properties:
        fields: { type: object, additionalProperties: true, description: "The ticket's fields (ADR 0031)." }
        id: { type: string }
        project_slug: { type: string }
        tracker_key: { type: string }
        title: { type: string }
        body_snapshot: { type: [string, "null"] }
        url: { type: [string, "null"] }
        size: { oneOf: [{ $ref: "#/components/schemas/Size" }, { type: "null" }] }
        risk: { oneOf: [{ $ref: "#/components/schemas/Risk" }, { type: "null" }] }
        state: { type: string }
        state_display: { type: [string, "null"] }
        workflow_name: { type: [string, "null"] }
        workflow_version: { type: [integer, "null"] }
        temporal_wf_id: { type: [string, "null"] }
        workflow_status:
          type: [string, "null"]
          description: The workflow's Temporal status, when requested and known. FAILED, TERMINATED or TIMED_OUT = the ticket is dead.
        failure:
          oneOf: [{ $ref: "#/components/schemas/WorkflowFailure" }, { type: "null" }]
        paused: { type: boolean, default: false }
        current_run: { oneOf: [{ $ref: "#/components/schemas/RunSummary" }, { type: "null" }] }
        pending_request: { oneOf: [{ $ref: "#/components/schemas/HumanRequest" }, { type: "null" }] }
        pr_url: { type: [string, "null"] }
        totals: { $ref: "#/components/schemas/Totals" }
        estimate: { oneOf: [{ $ref: "#/components/schemas/CostEstimate" }, { type: "null" }] }
        created_at: { type: string, format: date-time }
        closed_at: { type: [string, "null"], format: date-time }
    Totals:
      type: object
      properties:
        tokens_in: { type: integer, default: 0 }
        tokens_out: { type: integer, default: 0 }
        tokens_cached: { type: integer, default: 0 }
        cost_usd: { type: number, default: 0 }
        cost_eur: { type: number, default: 0 }
        duration_s: { type: number, default: 0 }
        runs: { type: integer, default: 0 }
    CostEstimate:
      type: object
      properties:
        median_usd: { type: [number, "null"] }
        p80_usd: { type: [number, "null"] }
        sample_size: { type: integer, default: 0 }
        over_p80: { type: boolean, default: false }
    WorkItemPage:
      type: object
      required: [items, meta]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/WorkItem" } }
        meta: { $ref: "#/components/schemas/PageMeta" }
    HandOff:
      type: object
      description: An output of a step and its receipts, in order (S25-04).
      required: [output, events]
      properties:
        output: { type: string, description: "spec_markdown, plan_markdown, or an output a transition declares" }
        current_digest: { type: [string, "null"], description: "The digest of what the output is worth today; null when it is gone." }
        events: { type: array, items: { $ref: "#/components/schemas/HandOffEvent" } }
    HandOffEvent:
      type: object
      description: A run produced or read an output, under this digest; a receipt never changes.
      required: [kind, digest, run_id, at]
      properties:
        kind: { type: string, enum: [produced, read] }
        digest: { type: string, description: "sha256:<hex> of the value's canonical JSON" }
        run_id: { type: string }
        stage: { type: [string, "null"] }
        attempt: { type: [integer, "null"] }
        at: { type: string, format: date-time }
    TimelineEntry:
      type: object
      required: [ts, kind, title]
      properties:
        ts: { type: string, format: date-time }
        kind: { type: string, enum: [state_change, run, decision, finding, release, comment, error] }
        title: { type: string }
        detail: { type: [string, "null"] }
        actor: { type: [string, "null"] }
        actor_kind: { type: string, enum: [user, agent, system], default: system }
        ref_id: { type: [string, "null"] }
        cost_usd: { type: [number, "null"] }
...
    WorkItemJourney:
      type: object
      description: >-
        A work item's path through its workflow (S22-01). `graph` and `process` describe the version
        the item is pinned to, not the project's current one; `moves` and `steps` are in time order.
      required: [work_item_id, state, graph, process, moves, steps]
      properties:
        work_item_id: { type: string }
        tracker_key: { type: [string, "null"] }
        workflow_name: { type: [string, "null"] }
        workflow_version: { type: [integer, "null"] }
        initial: { type: [string, "null"], description: The workflow's initial state. }
        state: { type: string, description: The state the item is in now. }
        closed: { type: boolean, default: false }
        graph: { $ref: "#/components/schemas/WorkflowGraph" }
        process: { type: array, items: { $ref: "#/components/schemas/ProcessStep" } }
        moves: { type: array, items: { $ref: "#/components/schemas/JourneyMove" } }
        steps: { type: array, items: { $ref: "#/components/schemas/JourneyStep" } }
...
    DecisionRequest:
      type: object
      required: [kind]
      properties:
        request_id: { type: [string, "null"] }
        kind: { type: string, enum: [approve, reject, answer, scope_change, complete] }
        answer: { type: [string, "null"] }
        reason: { type: [string, "null"] }
        granted_paths: { type: array, items: { type: string } }
        values: { type: object, additionalProperties: true, description: "`complete`: the values of the task's form, written into the ticket's fields." }
        attested: { type: boolean, default: false, description: "`complete`: the person attests the task's statement." }
    HumanRequest:
      type: object
      required: [id, kind, requested_at]
      properties:
        id: { type: string }
        work_item_id: { type: string }
        transition_id: { type: [string, "null"] }
        kind: { type: string, enum: [approval, question, scope_change, task] }
        payload: { type: object, additionalProperties: true }
        requested_at: { type: string, format: date-time }
        due_at: { type: [string, "null"], format: date-time }
        decided_by: { type: [string, "null"] }
        decided_at: { type: [string, "null"], format: date-time }
        decision: { type: [object, "null"], additionalProperties: true }
    WorkItemAction:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [pause, resume, stop, migrate, rerun_stage, mark_agent_ready] }
        workflow_def_id: { type: [string, "null"] }
        state_mapping: { type: object, additionalProperties: { type: string } }
        run_id: { type: [string, "null"] }
        reason: { type: [string, "null"] }
    RunSummary:
      type: object
      required: [id, status, stage_role]
      properties:
        id: { type: string }
        status: { type: string, enum: [queued, running, succeeded, failed, cancelled, timed_out] }
        stage_role: { type: string }
        attempt: { type: integer }
        backend: { type: [string, "null"] }
        model: { type: [string, "null"] }
        cost_usd: { type: number, default: 0 }
        started_at: { type: [string, "null"], format: date-time }
...